Fix stale console-sections references after sections.ts rename
The rename in 027d185bf missed AdminLayout.vue, routes.spec.ts, and the
suppression-admin-ui spec, which still imported/referenced the old
console-sections path. Also filters hidden sections out of the nav bands
and tightens the masthead block in AdminLayout.
